Add `makePixelPerfect`

+++ packages/figma-clean-document/src/clean-document/clean-layer.js
@@ -2,6 +2,7 @@ import { traverseLayer } from '@create-figma-plugin/utilities'
 import { smartSortChildLayers } from 'figma-sort-layers/src/smart-sort-child-layers'
 import { updateLayersSortOrder } from 'figma-sort-layers/src/update-layers-sort-order'
 import { deleteHiddenLayer } from '../delete-hidden-layers/delete-hidden-layer'
+import { makePixelPerfect } from '../make-pixel-perfect/make-pixel-perfect'
 import { smartRenameLayer } from '../smart-rename-layers/smart-rename-layer'
 import { ungroupSingleLayerGroup } from '../ungroup-single-layer-groups/ungroup-single-layer-group'
 
@@ -9,6 +10,7 @@ export function cleanLayer (
   layer,
   {
     deleteHiddenLayers,
+    pixelPerfect,
     smartRenameLayers,
     smartRenameLayersWhitelistRegex,
     smartSortLayers,
@@ -16,14 +18,23 @@ export function cleanLayer (
   }
 ) {
   traverseLayer(layer, function (layer) {
+    if (layer.type === 'PAGE' || layer.removed === true) {
+      return
+    }
     if (deleteHiddenLayers === true) {
       deleteHiddenLayer(layer)
+      if (layer.removed === true) {
+        return
+      }
     }
     if (ungroupSingleLayerGroups === true) {
       ungroupSingleLayerGroup(layer)
+      if (layer.removed === true) {
+        return
+      }
     }
-    if (layer.removed === true) {
-      return
+    if (pixelPerfect === true) {
+      makePixelPerfect(layer)
     }
     if (smartRenameLayers === true) {
       smartRenameLayer(layer, smartRenameLayersWhitelistRegex)
